Refactoring parser into two files
